make MCM reopen last mod config
small UX change that makes the MCM reopen the last mod config menu that the player opened.

This only takes effect if the user clicked on a mod config since launching the game, so the "Morrowind Script Extender" page is still displayed the first time the MCM is opened during each launch.

This is implemented using a variable that stores the name of the most recently opened mod config. I implemented the functionality this way in order to minimize the amount of code that had to be rewritten to add this feature.

leave loop early after finding mod