Rewrite the MCM menu for MWSE so that it uses modern syntax
The mod config menu for the `"Morrowind Script Extender"` was written using the outdated MCM syntax and the deprecated `mwse.registerMCM` function.

This commit updates the mod config menu so that it uses more modern syntax.

This does not result in any noticeable changes for the end-user (except that the `OnOffButton`s are now `YesNoButton`s.

Squashed in a commit that removes a redundant `class = "Info"`