Bugfix: Make `doOnce` compatible with `isRegistered` and `unregister`
Previously if `doOnce` was passed to `event.register`, then a new anonymous function was created, and then that function was registered. Here is a minimum code sample to reproduce the issue:
```lua
local function func()
    mwse.log("func callback")
end
event.register("load", func, {doOnce=true})
mwse.log("is func registered? %s", event.isRegistered("load", func))
```

This was not explained in the documentation and led to some weird behavior.

My solution was to create a conversion table that converts the `originalCallback` to the `newCallback`. (Where the `newCallback` is as it was before.) Whenever `event.unregister` is called on a `doOnce` function, the `originalCallback` is used to retrieve the `newCallback`, and then the relevant entry is purged from the conversions table, and the `newCallback` is unregistered.

The relevant entry from the conversion table is also purged from the `newCallback` itself, to minimize errors.