Add `addDefaultsToDescriptions`
This method will recursively add a string to the description of each `Setting` variable that displays the default value of that setting.

This works by using either `variable.defaultValue`, or by using a specified `defaultConfig`.

If called on a `Setting`, then the default description will be added

If called on a `Template`, `Page`, or `Category`, then the method will be propagated to all subcomponents.